+++ lib/squib/conf.rb
@@ -52,6 +52,7 @@ module Squib
         Squib::logger.info { "  using config: #{file}" }
         yaml = YAML.load_file(file) || {}
       end
+      warn_unrecognized(yaml)
       Conf.new(DEFAULTS.merge(yaml))
     end
 
@@ -111,6 +112,14 @@ module Squib
 
     def normalize_antialias
       @config_hash['antialias'] = ANTIALIAS_OPTS[@config_hash['antialias'].downcase.strip]
+    end
+
+    # Were there any unrecognized options in the config file?
+    def self.warn_unrecognized(yaml)
+      unrec = yaml.keys - DEFAULTS.keys
+      if unrec.any?
+        Squib::logger.warn "Unrecognized configuration option(s): #{unrec.join(',')}"
+      end
     end
 
   end